display inital card displays
Add componet for card displays and sub components that can be used
within it. Moved the ratio values into Vuex so they can more easily
be accessed globally in the app. Currently we are just hardcoding a
single card in the App component data, however this will be
configurable by the user in the future & have more default examples.

Install the required packages to use the typescript decoraters for Vuex.